Make test_word_precision_check() discoverable
