+++ jwst/tests/helpers.py
@@ -39,17 +39,3 @@ def word_precision_check(str1, str2, length=5):
     else:
         return True
     return False
-
-
-def test_word_precision_check():
-    """Test word_precision_check"""
-    s1 = "a b c"
-    s2 = "aa bb cc"
-    s3 = "aa bb cc dd"
-    s4 = "aazz bbzz cczz"
-
-    assert word_precision_check(s1, s1)
-    assert not word_precision_check(s1, s2)
-    assert word_precision_check(s1, s2, length=1)
-    assert not word_precision_check(s2, s3)
-    assert word_precision_check(s2, s4, length=2)
