+++ jwst/tests/test_infrastructure.py
@@ -1,0 +1,14 @@
+from .helpers import word_precision_check
+
+def test_word_precision_check():
+    """Test word_precision_check"""
+    s1 = "a b c"
+    s2 = "aa bb cc"
+    s3 = "aa bb cc dd"
+    s4 = "aazz bbzz cczz"
+
+    assert word_precision_check(s1, s1)
+    assert not word_precision_check(s1, s2)
+    assert word_precision_check(s1, s2, length=1)
+    assert not word_precision_check(s2, s3)
+    assert word_precision_check(s2, s4, length=2)
